requirements.txt
flask
opencv-python
python-dotenv
ultralytics